use parcel geometry from mogis lv03 for feature highlight
